调整时间单位

+++ BasicGameEngine/src/main/java/ykw/common/IGameLogic.java
@@ -6,7 +6,7 @@ public interface IGameLogic {
 
     void input(Window window);
 
-    void update(float interval);
+    void update(int interval);
 
     void render(Window window);
 }
